rewrite core algorithm and data structures
Fixes #2

Previously I made the incorrect assumption that gene_ids were
consecutive. They are consecutively generated, but a gene may get lost,
before it reaches the leaves of the coalescent. Thus, the total number
of genes was independent from the gene_ids.

In order to avoid this misconception, I had to switch to data structures
mapping gene_ids to genes by some other means than just the index
(eg. unordered_map). Also we now feature a set of valid gene_ids which
then is used for iteration in the output.

A similar change for gene_ids may follow in the future. As will
refactorings which will make the code simpler and more robust.

As this is a breaking change I bump the version number to 2.0.